Group collected content by language
Now the structure of `collections` is such that each collection is an
object with keys for each language, e.g. `collections.posts.en` for all
English posts.

I think I will reverse this so that the form is `collections.en.posts`.
That way I can define `collection` as `collections[lang]` at the lang
level. Hopefully that works.