Add send-email script
As part of this I made the JSON include the post's category.

Right now, the script just creates the Scheduled Send. I need to make it
send it as well.

Also need to add some styling or something to the email, probably via
staticcss and an emails-specific stylesheet.

Then I need to make it run in the GH Action workflow on the schedule.